Larger existing destination files not overwritten properly
When using cmsfscp to copy to an existing file, the contents
of the destination file are not truncated. Therefore, if
the destination file is larger than the source file, you
will have unwanted trailing data remaining in the file
after the copy. This one-liner adds the O_TRUNC flag to
the open() syscall to address this problem.